transceiver, uhd: output timestamps on monontonic error
The "Loss of monotonic time" error occurs when a timestamp
arrives from the device that is earlier in time than the
previous timestamp. The output is an ALERT level message
generally accompanied by a transmit side timeout from the
device.

UHD: Loss of monotonic time
UHD: Loss of monotonic time
UHD: Device send timed out

Add to the error description the timestamp values that
generated the error and output with ALERT rather than
ERR log.

Signed-off-by: Thomas Tsou <ttsou@vt.edu>

git-svn-id: http://wush.net/svn/range/software/public@3307 19bc5d8c-e614-43d4-8b26-e1612bc8e597